patch

+++ diff_analysis/src/main.rs
@@ -201,17 +201,24 @@ fn spawn_db_worker(
 fn delete_rows_after_compute(diffs: &Vec<DiffAnalysis>, conn: &mut DbConnectionInTransaction) {
     let mut pairs = String::new();
     for (i, diff) in diffs.iter().enumerate() {
-        if let DiffAnalysisJobResult::Diff(_) = diff.job_result {
+        if matches!(
+            diff.job_result,
+            DiffAnalysisJobResult::Diff(_) | DiffAnalysisJobResult::ErrTooManyFiles(_, _)
+        ) {
             pairs.push_str(&format!("({}, {})", diff.from_id, diff.to_id));
             if i != diffs.len() - 1 {
                 pairs.push_str(", ");
             }
         }
     }
-    // we have to delete (from_id, to_id) pairs, as alone they are not unique
-    let query = format!(
-        "DELETE FROM analysis.diffs_to_compute WHERE (from_id, to_id) IN ({})",
-        pairs
-    );
-    conn.execute(diesel::sql_query(query)).unwrap();
+
+    // we may have all Errs, in which case we don't need to delete anything
+    if !pairs.is_empty() {
+        // we have to delete (from_id, to_id) pairs, as alone they are not unique
+        let query = format!(
+            "DELETE FROM analysis.diffs_to_compute WHERE (from_id, to_id) IN ({})",
+            pairs
+        );
+        conn.execute(diesel::sql_query(query)).unwrap();
+    }
 }
